The ManyToOne Relation
symfony console make:entity
We need to update the Answer entity. Now, what should the new property be called... question_id? Actually, no. And this is one of the coolest, but trickiest things about Doctrine. Instead, call it simply question... because this property will hold an entire Question object... but more on that later.

symfony console make:migration