Rendering Answer Data & Saving Votes
"5 Minutes Ago" Strings

To convert the DateTime into a friendly string, we can install a nice bundle. At your terminal, run:
composer require "knplabs/knp-time-bundle:1.19.0"

Owning Vs Inverse Sides of a Relation
This is what I wanted to talk about. Each relation has two different sides and these sides have a name: the owning side and the inverse side. For a ManyToOne and OneToMany relationship, the owning side is always the ManyToOne side. And it's easy to remember: the owning side is where the foreign key column lives in the database. In this case, the answer table will have a question_id column so this is the "owning" side.

Inverse Side is Optional